Allow free allocation of gestures to bank slots
ringctl previously assumed that each gesture was used exactly once, in a
fixed slot in every bank. However, Neova actually allows free allocation
of gestures to slots, including duplicates and empty slots.

Rework the script, output format and command-line interface to support
this, and update the documentation to describe the new interface.